+++ src/test/scala/uk/gov/hmrc/test/ui/cucumber/stepdefs/TraderDetailsStepDef.scala
@@ -16,8 +16,11 @@
 
 package uk.gov.hmrc.test.ui.cucumber.stepdefs
 
-import uk.gov.hmrc.test.ui.pages.TraderDetails.TransitHolder.{AddHolderContactNamePage, AddHolderContactPersonPage, AddHolderContactPhoneNumberPage, HolderAddressPage, HolderEORINumberPage, IsHolderEORIKnownPage}
-import uk.gov.hmrc.test.ui.pages.TraderDetails.{ActingAsRepresentativePage, AddConsignorContactNamePage, AddConsignorContactPersonPage, AddConsignorContactPhoneNumberPage, ConsigneeAddressPage, ConsigneeEoriPage, ConsigneeNamePage, ConsignorAddressPage, ConsignorEoriPage, ConsignorNamePage, HolderNamePage, IsConsigneeEoriKnownPage, IsConsignorEoriKnownPage, IsTIRidKnownPage, MoreThanOneConsigneePage, RepresentativeCapacityPage, RepresentativeEORINumberPage, RepresentativeNamePage, RepresentativePhoneNumberPage, SetReducedDataSetIndicatorPage, TirHolderIdPage}
+import uk.gov.hmrc.test.ui.pages.TraderDetails.Consignee.{ConsigneeAddressPage, ConsigneeEoriPage, ConsigneeNamePage, IsConsigneeEoriKnownPage, MoreThanOneConsigneePage}
+import uk.gov.hmrc.test.ui.pages.TraderDetails.Consignor.{AddConsignorContactNamePage, AddConsignorContactPersonPage, AddConsignorContactPhoneNumberPage, ConsignorAddressPage, ConsignorEoriPage, ConsignorNamePage, IsConsignorEoriKnownPage}
+import uk.gov.hmrc.test.ui.pages.TraderDetails.Representative.{ActingAsRepresentativePage, RepresentativeCapacityPage, RepresentativeEORINumberPage, RepresentativeNamePage, RepresentativePhoneNumberPage}
+import uk.gov.hmrc.test.ui.pages.TraderDetails.TransitHolder.{AddHolderContactNamePage, AddHolderContactPersonPage, AddHolderContactPhoneNumberPage, HolderAddressPage, HolderEORINumberPage, HolderNamePage, IsHolderEORIKnownPage, IsTIRidKnownPage, TirHolderIdPage}
+import uk.gov.hmrc.test.ui.pages.TraderDetails.SetReducedDataSetIndicatorPage
 
 class TraderDetailsStepDef extends BaseStepDef {
